kubectl auth set when  cluster created by gcluster (#353)
