+++ src/xpk/commands/batch.py
@@ -18,7 +18,7 @@ from argparse import Namespace
 
 from ..core.kueue import LOCAL_QUEUE_NAME
 from ..utils.console import xpk_exit, xpk_print
-from .cluster import set_cluster_command
+from .common import set_cluster_command
 from ..core.core import add_zone_and_project
 from ..core.kjob import AppProfileDefaults
 from ..core.commands import run_command_for_value
