+++ src/xpk/commands/inspector.py
@@ -24,7 +24,7 @@ from ..core.core import (
 from ..core.kueue import CLUSTER_QUEUE_NAME, LOCAL_QUEUE_NAME
 from ..utils.file import append_tmp_file, write_tmp_file
 from ..utils.console import xpk_exit, xpk_print
-from .cluster import set_cluster_command
+from .common import set_cluster_command
 from .workload import get_workload_list
