+++ src/xpk/commands/job.py
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from .cluster import set_cluster_command
+from .common import set_cluster_command
 from .kind import set_local_cluster_command
 from ..core.commands import run_command_for_value, run_command_with_updates
 from ..utils.console import xpk_exit, xpk_print
